requirements.txt
aiogram==3.4.1
aiohttp==3.9.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
google-generativeai==0.3.2
python-dotenv==1.0.0